updateProject

+++ guowei-manager/guowei-manager-service/src/main/java/com/guowei/service/OrderService.java
@@ -27,7 +27,7 @@ public interface OrderService {
 	@Transactional(propagation=Propagation.NOT_SUPPORTED)
 	int addGwOrder(GwOrder order);
 	
-	int createGwOrder(GwOrder order, List<GwOrderdetail> orderdetail) throws UnsupportedEncodingException;
+	long createGwOrder(GwOrder order, List<GwOrderdetail> orderdetail) throws UnsupportedEncodingException;
 
 	int editGwOrder(GwOrder order);
